+++ packages/web3-redux/src/block/crud.ts
@@ -1,6 +1,6 @@
 import { name } from './common.js';
 import { BlockId, BlockTransaction, toPrimaryKey, validateId } from './model/index.js';
-import createCRUDModel from '../createCRUDModel.js';
+import createCRUDModel from '../crud/createCRUDModel.js';
 
 export const BlockCRUD = createCRUDModel<typeof name, BlockId, BlockTransaction, BlockTransaction>(name, {
     validateId,
