+++ packages/web3-redux/src/contracteventquery/crud.ts
@@ -7,7 +7,7 @@ import {
     validateId,
     toPrimaryKey,
 } from './model/index.js';
-import createCRUDModel from '../createCRUDModel.js';
+import createCRUDModel from '../crud/createCRUDModel.js';
 
 export const ContractEventQueryCRUD = createCRUDModel<
     typeof name,
