+++ packages/web3-redux/src/contractsend/crud.ts
@@ -1,6 +1,6 @@
 import { name } from './common.js';
 import { ContractSendId, ContractSend, validateId, validate } from './model/index.js';
-import createCRUDModel from '../createCRUDModel.js';
+import createCRUDModel from '../crud/createCRUDModel.js';
 
 export const ContractSendCRUD = createCRUDModel<typeof name, ContractSendId, ContractSend, ContractSend, ContractSend>(
     name,
