+++ packages/web3-redux/src/ethcall/crud.ts
@@ -1,6 +1,6 @@
 import { name } from './common.js';
 import { EthCallId, EthCall, validateId, validate, EthCallIndexInput, toPrimaryKey } from './model/index.js';
-import createCRUDModel from '../createCRUDModel.js';
+import createCRUDModel from '../crud/createCRUDModel.js';
 
 export const EthCallCRUD = createCRUDModel<typeof name, EthCallId, EthCall, EthCall, EthCallIndexInput>(name, {
     validateId,
