+++ packages/web3-redux/src/http/crud.ts
@@ -1,6 +1,6 @@
 import { name } from './common.js';
 import { HTTPCacheId, Http } from './model/index.js';
-import createCRUDModel from '../createCRUDModel.js';
+import createCRUDModel from '../crud/createCRUDModel.js';
 
 export const HTTPCacheCRUD = createCRUDModel<typeof name, HTTPCacheId, Http, Http, Http>(name);
 export default HTTPCacheCRUD;
