+++ packages/web3-redux/src/sync/crud.ts
@@ -1,6 +1,6 @@
 import { BaseSyncId, Sync, SyncIndexInput } from './model/index.js';
 import { name } from './common.js';
-import createCRUDModel from '../createCRUDModel.js';
+import createCRUDModel from '../crud/createCRUDModel.js';
 
 export const SyncCRUD = createCRUDModel<typeof name, BaseSyncId, Sync, Sync, SyncIndexInput>(name);
 export default SyncCRUD;
